pyproject.toml
[project]
name = "aistorytelling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kokoro-onnx>=0.2.6",
    "langchain-community>=0.3.14",
    "langchain-ollama>=0.2.2",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.62",
    "ollama>=0.4.5",
    "pydantic>=2.10.5",
    "soundfile>=0.13.0",
    "streamlit>=1.41.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
